Massive overhaul
This is a massive overhaul of the old index and added css to a different file in the same folder. The slideout js works and could be wired to a marker in mapbox. For some reason the site is cutoff on the right and the x10host site is acting up